# byclaw-tenant-node/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ByClaw fixed tenant node",
    "version": "1.0.0",
    "description": "BE-only mTLS API. Canonical business hash and Redis/Worker contracts are specified in README. No raw SQL/bootstrap/message overwrite API."
  },
  "servers": [
    {
      "url": "https://tenant-node.example.test:3100"
    }
  ],
  "security": [
    {
      "mtls": []
    }
  ],
  "paths": {
...
    "/internal/v1/schema": {
      "get": {
        "summary": "Observed and audited schema state",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/schema-tasks/{auditId}": {
      "get": {
        "summary": "Durable schema task result",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "name": "auditId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9:_-]{1,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResult"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/schema-tasks": {
      "post": {
        "summary": "Persist and accept an exact BE release bundle",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "$ref": "#/components/parameters/key"
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted; not execution success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["task", "bundle"],
                "properties": {
                  "task": {
                    "$ref": "#/components/schemas/SchemaTask"
                  },
                  "bundle": {
                    "type": "string",
                    "format": "binary",
                    "description": "ZIP <= 8 MiB; exact manifest and __ddl.sql entries"
                  }
                }
              },
              "encoding": {
                "task": {
                  "contentType": "application/json"
                },
                "bundle": {
                  "contentType": "application/zip"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/schema-tasks/{auditId}/report-ack": {
      "post": {
        "summary": "BE confirms its stored report",
        "parameters": [
          {
            "$ref": "#/components/parameters/enterprise"
          },
          {
            "$ref": "#/components/parameters/generation"
          },
          {
            "name": "auditId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9:_-]{1,64}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rejected; sanitized error code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["attemptNo", "status"],
                "properties": {
                  "attemptNo": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "status": {
                    "enum": ["VERIFIED", "FAILED", "NEEDS_ATTENTION"]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "mtls": {
        "type": "mutualTLS",
        "description": "Trusted BE certificate chain and pinned CN"
      }
    },
    "parameters": {
      "enterprise": {
        "name": "X-Enterprise-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
      "generation": {
        "name": "X-Tenant-Generation",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[1-9][0-9]{0,18}$",
          "description": "规范十进制 signed bigint ID；禁止 JS Number"
        }
      },
...
      "key": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "auditId:attemptNo"
      }
    },
    "schemas": {
...
      "SchemaScript": {
        "type": "object",
        "additionalProperties": false,
        "required": ["version", "parentVersion", "path", "sha256"],
        "properties": {
          "version": {
            "type": "string",
            "maxLength": 64
          },
          "parentVersion": {
            "type": ["string", "null"]
          },
          "path": {
            "type": "string",
            "pattern": "^(baseline|versions)/[A-Za-z0-9_.-]+/__ddl.sql$"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        }
      },
      "SchemaTask": {
        "type": "object",
        "required": [
          "protocolVersion",
          "enterpriseId",
          "generation",
          "dbSandboxRecordId",
          "auditId",
          "requestId",
          "attemptNo",
          "fencingToken",
          "operationType",
          "triggerType",
          "byclawReleaseVersion",
          "fromVersion",
          "targetVersion",
          "bundleDigest",
          "deadline",
          "scripts"
        ],
        "properties": {
          "protocolVersion": {
            "const": 1
          },
          "enterpriseId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "generation": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "dbSandboxRecordId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "规范十进制 signed bigint ID；禁止 JS Number"
          },
          "auditId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9:_-]{1,64}$"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9:_-]{1,64}$"
          },
          "attemptNo": {
            "type": "integer",
            "minimum": 1
          },
          "fencingToken": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "operationType": {
            "enum": ["INIT", "UPDATE"]
          },
          "triggerType": {
            "enum": ["MANUAL", "AUTO_PROVISION", "AUTO_RELEASE_UPGRADE"]
          },
          "byclawReleaseVersion": {
            "type": "string"
          },
          "fromVersion": {
            "type": ["string", "null"]
          },
          "targetVersion": {
            "type": "string"
          },
          "bundleDigest": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          },
          "scripts": {
            "type": "array",
            "minItems": 1,
            "maxItems": 32,
            "items": {
              "$ref": "#/components/schemas/SchemaScript"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code"],
            "properties": {
              "code": {
                "type": "string"
              }
            }
          }
        }
      },
      "SchemaResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SchemaTask"
          },
          {
            "type": "object",
            "required": ["status", "observedVersion", "cleanupStatus"],
            "properties": {
              "status": {
                "enum": [
                  "PENDING",
                  "RUNNING",
                  "VERIFYING",
                  "VERIFIED",
                  "FAILED",
                  "RECONCILING",
                  "NEEDS_ATTENTION"
                ]
              },
              "observedVersion": {
                "type": ["string", "null"]
              },
              "cleanupStatus": {
                "enum": ["RETAINED_UNTIL_ACK", "PENDING_RETRY", "DELETED"]
              },
              "failureStage": {
                "type": "string"
              },
              "failureScript": {
                "type": "string"
              },
              "errorCode": {
                "type": "string"
              },
              "sqlState": {
                "type": "string",
                "pattern": "^[A-Z0-9]{5}$"
              },
              "failureReason": {
                "type": "string"
              },
              "steps": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        ]
      }
    }
  }
}